Added changes to Receipt class, Payment Methods, and accessing it in the main application class to run.
git push origin AndrewMidterm
